backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
genanki
openai
pydantic
langchain
langchain-community
langchain-openai
langgraph
langsmith
python-dotenv
